requirements.txt
wheel
Pillow==8.2.0
Django==2.2.24
pytz==2021.1
djangorestframework==3.12.4
django-rest-auth==0.9.5
django-restql==0.11.2
django-filter==2.4.0
drf-yasg==1.20.0
drf-yasg[validation]
python-decouple==3.4